refactor: moved types to new models folder
